isolate embedded from decorators module
